+++ benchmark/arche/iterate/arche_test.go
@@ -103,6 +103,29 @@ func runArcheQuery(b *testing.B, count int) {
 	}
 }
 
+func runArcheQueryCached(b *testing.B, count int) {
+	b.StopTimer()
+	world := ecs.NewWorld()
+
+	posID := ecs.ComponentID[c.Position](&world)
+	rotID := ecs.ComponentID[c.Rotation](&world)
+
+	world.Batch().NewEntities(count, posID, rotID)
+
+	cf := world.Cache().Register(ecs.All(posID, rotID))
+	var filter ecs.Filter = &cf
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		query := world.Query(filter)
+		b.StartTimer()
+		for query.Next() {
+			pos := (*c.Position)(query.Get(posID))
+			pos.X = 1.0
+		}
+	}
+}
+
 func runArcheFilter(b *testing.B, count int) {
 	b.StopTimer()
 	world := ecs.NewWorld()
@@ -231,6 +254,43 @@ func runArcheQuery1kArch(b *testing.B, count int) {
 	}
 }
 
+func runArcheQuery1kArchCached(b *testing.B, count int) {
+	b.StopTimer()
+	world := ecs.NewWorld()
+	c.RegisterAll(&world)
+
+	perArch := 2 * count / 1000
+
+	for i := 0; i < 1024; i++ {
+		mask := i
+		add := make([]ecs.ID, 0, 10)
+		for j := 0; j < 10; j++ {
+			id := ecs.ID(j)
+			m := 1 << j
+			if mask&m == m {
+				add = append(add, id)
+			}
+		}
+		for j := 0; j < perArch; j++ {
+			entity := world.NewEntity()
+			world.Add(entity, add...)
+		}
+	}
+
+	cf := world.Cache().Register(ecs.All(6))
+	var filter ecs.Filter = &cf
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		query := world.Query(filter)
+		b.StartTimer()
+		for query.Next() {
+			pos := (*c.TestStruct6)(query.Get(6))
+			pos.Val = 1
+		}
+	}
+}
+
 func runArcheFilter1kArch(b *testing.B, count int) {
 	b.StopTimer()
 	world := ecs.NewWorld()
@@ -260,6 +320,79 @@ func runArcheFilter1kArch(b *testing.B, count int) {
 		b.StartTimer()
 		for query.Next() {
 			pos := (*c.TestStruct6)(query.Get(6))
+			pos.Val = 1
+		}
+	}
+}
+
+func runArcheQuery1Of1kArch(b *testing.B, count int) {
+	b.StopTimer()
+	world := ecs.NewWorld()
+	c.RegisterAll(&world)
+
+	perArch := 2 * count / 1000
+
+	for i := 0; i < 1024; i++ {
+		mask := i
+		add := make([]ecs.ID, 0, 10)
+		for j := 0; j < 10; j++ {
+			id := ecs.ID(j)
+			m := 1 << j
+			if mask&m == m {
+				add = append(add, id)
+			}
+		}
+		for j := 0; j < perArch; j++ {
+			entity := world.NewEntity()
+			world.Add(entity, add...)
+		}
+	}
+	world.Batch().NewEntities(count, 10)
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		query := world.Query(ecs.All(10))
+		b.StartTimer()
+		for query.Next() {
+			pos := (*c.TestStruct6)(query.Get(10))
+			pos.Val = 1
+		}
+	}
+}
+
+func runArcheQuery1Of1kArchCached(b *testing.B, count int) {
+	b.StopTimer()
+	world := ecs.NewWorld()
+	c.RegisterAll(&world)
+
+	perArch := 2 * count / 1000
+
+	for i := 0; i < 1024; i++ {
+		mask := i
+		add := make([]ecs.ID, 0, 10)
+		for j := 0; j < 10; j++ {
+			id := ecs.ID(j)
+			m := 1 << j
+			if mask&m == m {
+				add = append(add, id)
+			}
+		}
+		for j := 0; j < perArch; j++ {
+			entity := world.NewEntity()
+			world.Add(entity, add...)
+		}
+	}
+	world.Batch().NewEntities(count, 10)
+
+	cf := world.Cache().Register(ecs.All(10))
+	var filter ecs.Filter = &cf
+
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		query := world.Query(filter)
+		b.StartTimer()
+		for query.Next() {
+			pos := (*c.TestStruct10)(query.Get(10))
 			pos.Val = 1
 		}
 	}
@@ -357,6 +490,18 @@ func BenchmarkArcheIterQueryID_100_000(b *testing.B) {
 	runArcheQuery(b, 100000)
 }
 
+func BenchmarkArcheIterQueryIDCached_1_000(b *testing.B) {
+	runArcheQueryCached(b, 1000)
+}
+
+func BenchmarkArcheIterQueryIDCached_10_000(b *testing.B) {
+	runArcheQueryCached(b, 10000)
+}
+
+func BenchmarkArcheIterQueryIDCached_100_000(b *testing.B) {
+	runArcheQueryCached(b, 100000)
+}
+
 func BenchmarkArcheIterFilter_1_000(b *testing.B) {
 	runArcheFilter(b, 1000)
 }
@@ -441,6 +586,18 @@ func BenchmarkArcheIter1kArchID_100_000(b *testing.B) {
 	runArcheQuery1kArch(b, 100000)
 }
 
+func BenchmarkArcheIter1kArchIDCached_1_000(b *testing.B) {
+	runArcheQuery1kArchCached(b, 1000)
+}
+
+func BenchmarkArcheIter1kArchIDCached_10_000(b *testing.B) {
+	runArcheQuery1kArchCached(b, 10000)
+}
+
+func BenchmarkArcheIter1kArchIDCached_100_000(b *testing.B) {
+	runArcheQuery1kArchCached(b, 100000)
+}
+
 func BenchmarkArcheFilter1kArchID_1_000(b *testing.B) {
 	runArcheFilter1kArch(b, 1000)
 }
@@ -451,4 +608,28 @@ func BenchmarkArcheFilter1kArchID_10_000(b *testing.B) {
 
 func BenchmarkArcheFilter1kArchID_100_000(b *testing.B) {
 	runArcheFilter1kArch(b, 100000)
+}
+
+func BenchmarkArcheIter1Of1kArch_1_000(b *testing.B) {
+	runArcheQuery1Of1kArch(b, 1000)
+}
+
+func BenchmarkArcheIter1Of1kArch_10_000(b *testing.B) {
+	runArcheQuery1Of1kArch(b, 10000)
+}
+
+func BenchmarkArcheIter1Of1kArch_100_000(b *testing.B) {
+	runArcheQuery1Of1kArch(b, 100000)
+}
+
+func BenchmarkArcheIter1Of1kArchCached_1_000(b *testing.B) {
+	runArcheQuery1Of1kArchCached(b, 1000)
+}
+
+func BenchmarkArcheIter1Of1kArchCached_10_000(b *testing.B) {
+	runArcheQuery1Of1kArchCached(b, 10000)
+}
+
+func BenchmarkArcheIter1Of1kArchCached_100_000(b *testing.B) {
+	runArcheQuery1Of1kArchCached(b, 100000)
 }