+++ ecs/paged.go
@@ -31,3 +31,33 @@ func (p *pagedArr32[T]) Get(index int) *T {
 func (p *pagedArr32[T]) Len() int {
 	return p.len
 }
+
+// pagedPointerArr32 is a paged collection working with pages of length 32 arrays.
+//
+// Implements [archetypes].
+type pagedPointerArr32[T any] struct {
+	pages   [][fixedPageSize]*T
+	len     int
+	lenLast int
+}
+
+// Add adds a value to the paged array.
+func (p *pagedPointerArr32[T]) Add(value *T) {
+	if p.len == 0 || p.lenLast == fixedPageSize {
+		p.pages = append(p.pages, [fixedPageSize]*T{})
+		p.lenLast = 0
+	}
+	p.pages[len(p.pages)-1][p.lenLast] = value
+	p.len++
+	p.lenLast++
+}
+
+// Get returns the value at the given index.
+func (p *pagedPointerArr32[T]) Get(index int) *T {
+	return p.pages[index/fixedPageSize][index%fixedPageSize]
+}
+
+// Len returns the current number of items in the paged array.
+func (p *pagedPointerArr32[T]) Len() int {
+	return p.len
+}